Add maps visibility to flood-drought and oil-spill
This commit replaces the placeholders put in place for the Google Maps
API with the actual JS component. Because the API key needs to be exposed
client-side, we have to request it from the backend. The current solution
involves a hard-coded setup for local development, but changes will need
to be made regarding how the key is retrieved and how the frontend
communicates with the backend for production.